Introduce `unvme-cli`
`unvme-cli` is a command line interface utility to test PCIe-based NVMe
controller without linux kernel driver intervention.  This is helpful to
introduce user-defined scenarios which are especially not possible to
happen by the kernel driver's behaviors.

`unvme-cli` is working with `unvmed` which stands for unvme daemon
process to keep the device resource information.  `unvme` program will
talk to the `unvmed` daemon by requesting actions for daemon to do them.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>