unvmed: Allow show-regs without enabling NVMe
show-regs is nothing but a command to print out controller property
registers located in BAR0 mmio memory space.  This has nothing to do
with enabling NVMe controller (CC.EN=1) so that this patch removed the
condition that NVMe should be enabled for this command.

Due to the previous patch, `unvme add <bdf>` is enough to run `unvme
show-regs` command.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>