unvmed: Remove `enable-ctrl` command
`unvme enable-ctrl <bdf>` command has requested `libvfn` to initialize
NVMe controller along with PCI initializztion.  This was too much for
this application which wants to enable the controller and that's it.  No
admin commands are to be issued.

From now on, `unvme enable <bdf>` will replace the previous enable-ctrl
command.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>