unvmed: Manage sq/cq pairs in unvme instance
`libvfn` manages SQ/CQ instances based on dynamic array.  But in
`unvme-cli`, we support to create I/O queue with a specific qid which
means some of SQ/CQ array can be NULL.  To prevent crash when `libvfn`
tries to free the instances in the array, we should manage all the
created queue instances in this program.

This patch manages the sq/cq instances based on the list and when nvme
controller is reset, queue resources are going to be freed by
__free_nvme().

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>