unvmed: Set nsqa and ncqa during initialization
Since unvme-cli does not follow all the behaviors of the driver, it does
not set up nsqa and ncqa.
Originally, nsqa and ncqa are set through the set feature command, but
we cannot expect the user to send the set feature command.
Therefore, during the `unvme_add`, nsqa and ncqa are set respectively to
nsqr and ncqr.

Signed-off-by: Jungwon Lee <jjung1.lee@samsung.com>