unvmed: Add '--nr-ioqs' option to `unvme add`
When adding a new NVMe PCI device to unvmed, it has initialized `libvfn`
with default option parameter (i.e., struct nvme_ctrl_opts) by giving
NULL to nvme_ctrl_init.  The defualt number of queue to be requested to
the device is 63 (zero-based) in `libvfn`.

This might cause segfault when unvme-cli calls nvme_close() when `unvme
del` command to detach it from unvmed since `libvfn` will clean up the
queue instnaces and memory based on the ctrl.opts.nsqr and ncqr.

This patch explicitly configures the ctrl.opts.nsqr and ncqr by giving
it to nvme_ctrl_init and this can be received by users as option (-q or
--nr-ioqs).

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>