unvmed: Call close vfio instance in `unvme del`
`unvme del` command is to detach the target device from unvmed and close
the opened file descriptors along with the mmaped resources through
vfio.  nvme_close in `libvfn` is the one we should call from `unvme del`
command since it's going to free all the nvme-realted queue resources
along with the vfio instances.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>