unvmed: mkdir /var/log for creating log file
If the /var/log directory does not exist and an attempt is made to
create an unvme.log file in that directory, an error occurs.
To prevent this, add a command to make the /var/log directory, and if an
error occurs except for when the directory already exists, terminate the
process.

Signed-off-by: Jungwon Lee <jjung1.lee@samsung.com>