nvmed: Add `unvme_sq_exists` function
The existing `unvme_sq` function returns the address of the sq[qid], and
libvfn does not guarantee if a sq[qid] has actually been created by the
address since it's just an array in libvfn.
The `unvme_sq_exists` searches in the list of SQs added when creating SQ
and returns whether sq[qid] has actually been created.

Signed-off-by: Jungwon Lee <jjung1.lee@samsung.com>