unvmed: Return error msg to client in signal
When unexpected signals, such as SIGSEGV or SIGABRT occurs, daemon
process will be terminated unexpectedly.  In this case, client will keep
waiting for the daemon process to send a message through the msgq, which
means client will never know the daemon has been terminated or not.

This patch added signal handlers for those two signals and if they
happened, it will send the error msg back to the client with -ENOTCONN
errno.  Then, client will show the error message and terminated with
errno.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>