unvmed: introduce 'libunvmed' for driver behavior
Introduce 'libunvmed' which stands for unvme driver or daemon library
to be shared with external tools such as fio which runs on the unvmed
(to be supported soon).  Along with this, this patch also added a
dependency of 'libnvme >= 1.8.0' to re-use NVMe spec-defined struct and
enumerations rather than using hard-coded values.  It's not linked yet
since this patch just used enum and struct values and some of macros
defined in the header files in 'libnvme'.

This patch is a pretty huge update to separate driver behaviors from the
unvmed-*.c to lib/libunvmed.c.  Also it renamed unvmed_* functions in
unvmed_*.c to avoid confusions between libunvmed and unvmed-*.c. No
functional changes were made.

This aptch also added lib/ directory with COPYING and LICENSE file for
the dual-licenses: LGPL-2.1-or-later and MIT.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>