unvme: send msg from daemon to client based on pid
The very first member of `struct unvme_msg` type represents msgq type to
be distinguished among multiple messages.  This can be used as an
identifier saying where to retrieve a message from the given msgq.

If type == 0 is given, msgrcv() system call tries to fetch the first
queue entry from the msgq regardless to the type in the message.  If
type > 0, msgrcv() retrieves the first msg whose queue type is type.

unvmed fetches one or more client messages from the msgq based on the
type as a pid.  Clients send their own msg to the msgq with type as
their pid.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>