unvme: don't pass unvme_msg to unvmed-file
unvmed-file APIs should just read or write the given 'filename' in the
current pwd.  But, daemon process can't simply determine the current
working directory (cwd) of the cli process, so it should always play
around with absolute path of the target file to read or write.

We used to give struct unvme_msg to determine the current working
directory in unvmed-file, but this should be taken care in command
handler rather than file handlers.

This patch removed unvme_msg dependencies and removed the header include
unvme.h.  Give the responsibility to give the absolute path of the
target file to the command handlers.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>