unvme: fix '-n|--nvme' filtering in log command
If we use 'read' system call in reading log file, it may fail in
filtering prefix which starts in a line since read() reads data based on
the given buffer rather than a 'line' of text.  Instead of using read()
for the text log file, this patch used getline() which returns allocated
string of line with number of bytes.  This makes us filter the very
first prefix like 'NVME' based on line by line.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>