unvme: manage ns instances and print in 'unvme status'
Added print of namespace instances registered to the current controller
instance.  unvme-cli is a tool communicating with the device as user
requested, which means it should not issue identify namespace command by
its own decision.  This patch added an inaterface to let user request to
unvmed to initialize and register the namespace instance to be managed
in the controller by '--init' option in 'unvme id-ns' command.

If user issues 'unvme id-ns <bdf> -n <nsid> --init', it will register
the namespace instance (struct unvme_ns) to the controller and it's
printed in 'unvme status' command.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>