unvme-cli: use provided buffer rather than a new one
unvmed_alloc_cmd() used to allocate a user data buffer inside of it
which caused double buffer copy from this buffer to the user buffer
which is given by the application.

This patch added an argument 'buf' to unvmed_alloc_cmd() and this helper
maps the given buffer to the IOMMU to make device DMA from or to the
given buffer directly.  To do this, this patch also removed all the
malloc() from the unmved-cmds.c and converted it to aligned_alloc() to
make sure the allocated memory is aligned to the page size.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>