unvme: limit only one app command running at a time
UNVME_APP_CMD means this command executes an external tool like fio in
the current process context as a thread.  It's difficult to redirect the
stdout and stderr used in the external tools to a file because all the
threads in the current daemon process share the same stdout and stderr
file stream.  So, the process's stdout and stderr should be retrieved by
a single client at a time.

This patch added UNVME_APP_CMD and make sure that only a single app
command is running at a time to avoid stdout/stderr corruption.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>