unvme: add support signal handling for a job
This patch added a concept named unvme_job and it represents a single
instance running in the daemon process context.  Everytime client CLI
process requests a job to the daemon process, it will add the instance
to a linked list.

With this, added support for signal handling for a job running in the
daemon process context.  For example, if we run fio application in the
daemon process context, cli now can signal the running job to the daemon
process.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>