unvme: check queue existence
I/O SQ, CQ-related commands have been designed to issue corresponding
NVMe admin commands to the controller, but queue context should also be
initialied to libvfn so that we can keep the queue context to libunvmed
driver.

For now, libvfn totally manages the sq and cq instances and we don't
want to make any confusions to the existing commands.  This patch
limited issuing the actual admin commands to controller if it's invalid
cases to avoid unexpected behaviors in libvfn.

Once libvfn exports nvme_configure_sq and nvme_configure_cq APIs to
public, we can separate the queue setup and issuing commands parts
separately so that user can choose to force issuing the command or not.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>