unvmed: manage unvme_sq, unvme_cq for reset handling
Added unvme queue managing to handle reset situation under fio I/O
workloads.  This patch introduced new structures: unvme_sq and unvme_cq.
Ealier, we've used nvme_sq and nvme_cq provided by libvfn directly, but
in that case reset handler cannot grab the mutual exclusive lock for the
corresponding queues to cancel the inflight commands.  To manage queue
instance ownership, added usq and ucq instance.

These sq cq queue context instances are not going to be freed up until
the device is deleted from the daemon process, and they are going to be
re-used in the corresponding session.  This is because if we free up the
queue instance itself when reset (e.g., controller reset) occurs,
application in uppser layer might not be aware the reset situation down
there and it might go through the queue to put commands and it might
cause NULL pointer dereferences.

lib/libunvmed.h defines the two instance struct and lib/libunvmed.c
defines its own private structure with the same struct in the beginning
of the struct to inter-changably use between the two types.  Private
instance has linked list node to be managed inside of the unvmed daemon.
This patch implemented the private instance started with '__' and it
should be used inside of the daemon process only, not being exported to
the public APIs.  And private struct also has the same struct type as an
anonymous struct.

app/fio should always be running jobs on each submission queue which
means there are no requirements to have locks at all.  But, to sync-up
the queue context between normal and reset situation, this patch added
lock concept to each submission queue with spinlock.  This spinlock will
not race for sure between I/O workers, but it should race with reset
handlers or sometimes with CLI client processes.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>