unvme: emit log when __msg is non-NULL
When unvmed daemon service is about to terminated, it emits a log
saying it's being terminated.  void exit(int) function is nothing but a
just overrided function which will be called in the daemon job thread
when the application ends up.  In that case, __msg might be NULL and it
causes segfault when emitting the log.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>