unvme: terminate and kill unvme* all processes
'unvme stop' command is to terminate all processes related to unvme and
unvmed daemon service running in the current system.  Ealier, it used to
terminate gracefully the unvmed daemon process only, but in abnormal
cases, client process might be in stuck situation and unvmed daemon
process also.  In this case, we should kill them all to restart the
daemon process.

This patch first tries to termiates processes named 'unvme*' and if they
are not terminated successfully, kill them all with SIGKILL.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>