unvme: add -a|--all option to 'unvme stop'
'unvme stop' used to terminate unvmed daemon service running in the
background.  And recently, it's been used to terminate all the client
processes along with the daemon process.

But, in most cases, we should terminate the daemon process so that
client processes just can wake up due to error of msgq.  But, in some of
error cases like fio hang situation, client process will keep waiting
for the response.  In this case, we can turn off the client processes as
well with -a|--all option.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>