unvme: set default fio shared object name to 'fio.so'
'fio' is actually not a shared object, it's a tool itself.  But,
unvme-cli uses this tool as a shared object to dynamically load the code
and data to the current process context and unload it once fio is done.

This patch decided the default shared object name of fio.  If user
does not give the --with-fio=<path> option to 'unvme start' command, it
will find out the default sahred object named 'fio.so' in the current
system.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>