fio: libunvmed: check if iomem exists
A fio job allocates xfer_buf at once at the initial time to manage data
buffer within a single memory buffer.  That buffer is stored in
td->orig_buffer.  libunvmed ioengine keeps the size of the buffer in
ld->orig_buffer_size.

In case of TRIM-only workloads (e.g., --rw=trim), td->orig_buffer is not
allocated at the initial time since it's not a data-based transfer even
though it requires PRP to form the LBA ranges.  To avoid zero-sized
memory mapping with IOMMU, this patch added checks if the orig_buffer
exists or not.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>