fio: libunvmed: add support TRIM command
This patch added support for asynchronous TRIM command in libunvmed
ioengine.  The only differnce between normal READ/WRITE and TRIM command
is that it allocates the PRP buffer and map the buffer to IOMMU
explicitly before sumitting the command to the device.  And free up the
PRP buffer in .event() callback after complete the command handling.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>