unvmed: fix non-aligned identify buffer
unvmed_init_ctrl() issues an Identify Namespace admin command if
'identify' argument is not given.  Previously it's been giving struct
nvme_id_ns local variable pointer to be mapped to the IOMMU, but it
might not be aligned to the page size.  It will fail in mapping the
vaddr to IOMMU iova.

This patch pgmap() for the identify data structure to get a aligned
memory buffer.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>