unvme: fix valgrind error for strncpy
'char bdf[13]' is nothing but an 13-bytes array and it should not be
copied with strncpy which considers the NULL terminated character.
unvme_check_args() prepares the bdf with sprintf so that '\0' terminated
charater will be there in the array.  So, we just can copy as a memory,
not string to avoid valgrind error.

==4168== Conditional jump or move depends on uninitialised value(s)
==4168==    at 0x484F02A: strncpy (in /usr/libexec/valgrind/vgpreload_memcheck-amd64-linux.so)
==4168==    by 0x10FDD2: main (unvme.c:442)
==4168==  Uninitialised value was created by a stack allocation
==4168==    at 0x10FC7F: main (unvme.c:422)

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>