unvme-cli: pass allocated length of user buffer
libunvmed should give buffer with a given size to unmap(free) the
buffer properly.  libunvmed manages the buffer with `cmd->vaddr` and
`cmd->len`.  Required data size and allocated buffer size might be
different so that application should pass the actual allocated size of
buffer to the library to free up them properly.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>
Reviewed-by: Jonghwi Jeong <jongh2.jeong@samsung.com>
Reviewed-by: Jungwon Lee <jjung1.lee@samsung.com>