libunvmed: remove unvmed_sq_update_tail_and_wait()
Remove unvmed_sq_update_tail_and_wait() since libunvmed has already
provided APIs to update the tail doorbell and wait for the corresponding
CQ entries to be posted by the device.  Updating and waiting is a
respnosibility for upper layer, not the library itself.

Remove the high-level API and let application do the same thing.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>
Reviewed-by: Jonghwi Jeong <jongh2.jeong@samsung.com>