libunvmed: fix unmapped leaked IOMMU mapping id_ns
unvmed_init_ns() API support to issue an Identify Namespace admin
command to the controller if user does not provide @identify data to
this API.   While doing 'unvme reset <bdf> --reinit', unvmed_init_ns()
is called without @identify data provided so that it issues the admin
command to the controller.

In this case, identify user data buffer is allocated by pgmap() and it's
mapped to the IOMMU table inside of unvmed_alloc_cmd() and the mapping
will be unmapped when unvmed_cmd_free() is explicitly called by the
caller, which is it.  If we don't free up the command instance properly
here, the VA <-> IOVA mapping relationship will be alive after this
function, and it's going to corrupt the newly allocated memory, which is
the same buffer of the identify data.

We *MUST* unmap the IOMMU relationship along with free up the data
buffer by pgunmap().  This patch added unvmed_cmd_free() call in
unvmed_init_ns() to make data buffer and IOMMU table consistent.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>