libunvmed: LOAD latest value for queue and cmd attributes
Head pointer and phase bit value are updated in another thread context
which might be the application upper layer and the cancelation thread
context might not fetch the latest value of the head pointer.  This
patch guaranteed to load (READ_ONCE) the head and phase value before
starting the cancelation.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>