fio: clean up mess in unexpected exit() call
unvmed loads fio symbols dynamically in the runtime and free it up with
dlclose() once 'fio' returns from main() as unvmed_run_fio() does.  But,
we can't simply guarantee that fio will never call exit() inside of it
since 'fio' itself is a standalone application which is not made for
shared object, so yes, it can definitely exit().

For example, fio calls exit() if parsing command line arguments fails.
In this case, the pthread where the fio context is running on will be
terminated without returning to the caller in unvmed_run_fio() and the
following clean up codes will never be executed.  If we don't clean up
the mess properly, the second run, especially when the first run was
fail with exit() call, will fail due to non-initialized global
variables.

To fix this, call setjmp() before calling main() of fio to make sure to
get back to here when fio calls exit() which leads to our own exit()
symbol defined in unvmed-cmds.c.  and unvmed will check the thread
variable __jump is set or not, if so, it longjmp() back to here with
return value EINTR and it goes to out label to clean up all the mess.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>
Reviewed-by: Jonghwi Jeong <jongh2.jeong@samsung.com>