unvme-cli: add ``unvme set-features`` command
Set Features admin command is a mandatory command to issue during the
initialization phase of a controller with features such as "Number of
Queues" to request a number of I/O queues to be allocated by the
controller.  Some controllers does not allocate I/O queue resources
within the device controller if host does not issue Set Features to
request the resources.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>