fio: add ``--thread=1`` by default to cmdline
Even though ``--thread=1`` is not a command-line option in fio, which is
a job option, add this one by default regardless to that whether user
gave ``--thread=1`` parameter or not.

There is a limitation on this default option.  If a jobfile is given,
and if the jobfile has explicit ``thread=0`` inside of it, jobfile will
overwrite the default option and it will fail with an error message.

Additionally, even if user gives ``--thread=0`` explicitly in the
command line interface, it will be ignored since ``--thread=1`` default
option will be appended to the end of the command line.

Able to run:
	unvme fio --name=test --ioengine=libunvmed \
		--filename=0000.01.00.0 --nsid=1 \
		--rw=read --bs=128k \
		--thread=0  (--thread=1)

Not able to run:
	[global]
	ioengine=libunvmed
	filename=0000.01.00.0
	nsid=1

	[job]
	rw=read
	bs=128k
	number_ios=1024
	thread=0

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>