fio/libunvmed: fix handle unsupported command properly
If fio tries to queue some commands unsupported by engine, libunvmed-
engine returns `-ENOTSUP` directly in `queue` function.  However, fio
cannot catch this error and considers this command has been submitted.
Because of this, even though the command is not submitted, fio tries to
reap the CQ entry of the command, and will be stucked in the while loop.

This patch sets `ENOTSUP` to `io_u->error` and returns 0 for fio to
spit some errors and stop submitting remaining commands.

Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>