meson: support --with-libvfn= with header included
`meson setup <builddir> --with-libvfn=<path>` targets a specific
location of libvfn installed which is not in the system library path
(e.g., /lib/, /usr/lib/).  This is useful while developing this tool
with libvfn under being developed.

To prepare a specific version of libvfn, user should build the libvfn
and install the libvfn to a specific location with:

	cd libvfn
	meson setup <builddir> --prefix=<path> --libdir=lib
	meson install -C <builddir>

	cd unvme-cli
	meson setup <builddir> --with-libvfn=<path>
	meson install -C <builddir>

The <path> will look like below:

	bin/
		vfntool
	include/
		vfn/**/*.h
	lib/
		libvfn.so*

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>