fio/libunvmed: use @io_u->buf in verifying TRIM-ed area
--trim_backlog=N option in fio represents the number of WRITE commands
before issuing a TRIM command.  --trim_percentage=% represents the
probability of whether to issue TRIM or not when the time comes.  In
this case, fio prepares @io_u instances for TRIM commands in
get_next_trim() where @io_u->xfer_buf is set to NULL and
@io_u->xfer_buflen > 0.  In NVMe, however, we definitely needs data
buffer for DSM command to represent the LBA ranges to discard.

This patch elaborates more details in @ld->trim_iomem, and use
@io_u->buf rather than @io_u->xfer_buf which is NULL when manipulating
DSM command.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>