unvmed-cmds: set default UUID for VFTOKEN
The kernel requires a VF token to open a device fd. Therefore, libvfn
expects the VFTOKEN environment variable to be set in
`vfio_get_device_fd`. To support this, we set a default UUID in
unvme-cli if the user does not provide one. This ensures compatibility
while still allowing overrides via --vf-token.

Signed-off-by: Haeun Kim <hanee.kim@samsung.com>