fio/libunvmed: update tail doorbell in @usq is disabled
@usq can be disabled anytime, but the instance can be alive.  For
example, in graceful controller reset sequence, libunvmed will disable
and quiesce submission queue to finish the in-flight commands
gracefully.  In this case, to guarantee to update the tail doorbell of
SQ, libunvmed ioengine should check whether this is really under deleted
queue or not.  By checking the @usq->refcnt > 1, which means libunvmed
still manages the @usq alive, we can update the tail doorbell.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>