unvme-cli: add support async commands
Previosly, we've guaranteed for the CQ entry to be fetched while
grabbing @usq lock to make sure other CQ entries not to be jumbled when
the caller is waiting for the command to complete.

This patch added support for asynchronous read/write cli commands by
adding two more fields in &struct unvme_cmd: completed and cqe
respectively.  @cmd->completed indicates whether the corresponding @cmd
instance is completed or not.  This field is set by the RCQ thread in
the background when the corresponding CQ entry is fetched by the CQ,
which means this field is only set in case interrupt is enabled (e.g.,
MSI-X) to enable the RCQ thread in the background.  Along with this,
@cmd->cqe will be updated when the CQ entry is fetched from the CQ and
the entry will not be pushed to the RCQ since it will be handled by the
caller directly.  To make this happen, caller should give
UNVMED_CMD_F_WAKEUP_ON_CQE to @cmd->flags.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>