unvmed: prevent --data-len= file I/O for user DPTRs
If user gives --data-len > 0 in `unvme passthru` command, libunvmed will
allocate a DMA buffer for the given --data-size and map it to IOMMU.  It
guarantees that the DMA buffer will be mapped to a contiguous CPU
virtual address which means unvmed easily access to that memory.  In
this case, file I/O can be done simply.

Let's say that user gives --prp1 != 0 or --prp2 != 0, one of the
following cases:

user wants to
	- inject invalid PRP address to test fault
	- put user-defined DPTR IOVAs allocated by `unvme malloc`

For both cases, passthru can't easily do file I/O since the given IOVAs
might not be mapped into IOMMU or not.  Therefore, this patch prevents
providing --prp1= or --prp2= with --data-len.  In short, If --data-len
is given, it means user-defined PRPs cannot be given.  Instead, if user
wants to give own PRP address and wants to read or write the DMA buffer
data, `unvme mread` or `unvme mwrite` should be used.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>