unvmed/cmds: reset controller if already enabled
If the actual NVMe device itself has already been enabled, which means
CSTS.RDY = 1 with CC.EN = 1, it's going to make consistency mismatch
between the @u->state flags and the actual state of the
controller(enabled).

`unvme add` command is an ENTRY command for unvmed to add the given PCI
device and go with unvmed from the beginning, which means if `unvme add`
command completes successfully, it *must* ensure that the device itself
is consistent with the @u instance managed in libunvmed.

Therefore, reset the controller if CSTS.RDY is already been set by the
device.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>