libunvmed: fix return EILSEQ in fail to delete CQ
If `delete-iocq` command is submitted, it might be failed with `Command
Sequence Error` in the device, if S/Q deletion is not proceeded first.
This is obviously an error case, but `unvme-cli` handles this as an
success case, because `errno` is not set.  To handle this as an error,
we set the @errno to EILSEQ.

Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>