unvme: remove app stdio handler
app, which represents an external app such as fio without modification,
typically considers stdout and stderr is ready, but in our use cases,
stdout and stderr should be redirected to the client process's one.

This patch redirected unvmed daemon process stdout and stderr to the
requested client stdout and stderr temporarily and reset them back to
the file located in /var/log/unvmed.stdout and .stderr.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>