unvme: remove waitpid and wait for pid file removed
waitpid() should have not been used to terminate the daemon process
since the `unvme stop` client process is not a parent process of the
daemon process since it has never forked the daemon at all, the one who
forked before is `unvme start` client process.  waitpid() is just to
wait for the *child* processes of the current process.

Remove waitpid() since it caused direct return from waitpid() with
negative error and it kills the daemon process with SIGKILL even if it's
still on-going on termination.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>