unvme-cli: add support -c|--cid in `unvme passthru`
In the previous patch, we started supporting CID mgmt. for commands.
This patch added support -c|--cid argument in `unvme passthru` command
which is pretty much flexible command for users to manipulate command
structure.  If -c|--cid is given within a range of [0, qsize), it will
_try_ to allocate the given specific cid to libunvmed, and it might fail
in case that libunvmed has already allocated the specific cid value,
otherwise it will be successfully done.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>